Move scroll event to mouse events
