+++ resin/resin/core/window.cpp
@@ -14,7 +14,6 @@
 #include <resin/event/event.hpp>
 #include <resin/event/key_events.hpp>
 #include <resin/event/mouse_events.hpp>
-#include <resin/event/scroll_events.hpp>
 #include <resin/event/window_events.hpp>
 #include <resin/imgui/gizmo.hpp>
 #include <string>
