+++ resin/resin/resin.hpp
@@ -20,7 +20,6 @@
 #include <resin/event/event.hpp>
 #include <resin/event/key_events.hpp>
 #include <resin/event/mouse_events.hpp>
-#include <resin/event/scroll_events.hpp>
 #include <resin/event/window_events.hpp>
 #include <resin/imgui/gizmo.hpp>
 #include <resin/resources/resource_managers.hpp>
